requirements.txt
asgiref==3.2.3
Django==3.0.3
django-sql==1.0.1
djangorestframework==3.11.0
Pillow==7.1.0
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.0
